add binder, tweak metadata and exports
