+++ .binder/environment.yml
@@ -1,0 +1,16 @@
+name: jupyterlab_kaitai
+
+channels:
+  - conda-forge
+
+dependencies:
+  - ipywidgets >=7,<8
+  - jupyter-packaging >=0.7.9,<0.8
+  - jupyterlab =3
+  - nodejs >=12,!=13.*,<15
+  - numpy
+  - pip
+  - python >=3.6
+  - wheel
+  - pip:
+    - kaitaistruct
